Complete changeover to BigDecimal-based coordinate converter.
Tests, etc.
