+++ objectRecognitionSource/src/main/java/io/vantiq/extsrc/objectRecognition/imageRetriever/CoordinateConverter.java
@@ -8,16 +8,12 @@
 
 package io.vantiq.extsrc.objectRecognition.imageRetriever;
 
-import static org.opencv.core.CvType.CV_32FC2;
-import static org.opencv.core.CvType.CV_64FC1;
-
-import org.opencv.core.Core;
-import org.opencv.core.Mat;
-import org.opencv.core.Size;
-import org.opencv.imgproc.Imgproc;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.math.BigDecimal;
+import java.math.MathContext;
+import java.math.RoundingMode;
 
 /**
  * Construct a coordinate converter based on the 4 non-collinear points provided.  This converter
@@ -39,9 +35,10 @@ public class CoordinateConverter {
     @SuppressWarnings({"WeakerAccess"})
     Logger log = LoggerFactory.getLogger(this.getClass());
 
-    private Mat converter = null;
-    private Mat src = null;
-    private Mat dst = null;
+    private BigDecimal[] converter = null;
+    private BigDecimal[] src = null;
+    private BigDecimal[] dst = null;
+    private MathContext mc = new MathContext(8, RoundingMode.CEILING);
 
 
     /**
@@ -62,11 +59,11 @@ public class CoordinateConverter {
      * @param source 2D array of floats defining the source space.  Must be of length 4
      * @param destination 2D array of floats defining the target space.  Length 4, where each point corresponds to those in src.
      */
-    public CoordinateConverter(Float[][] source, Float[][] destination) {
+    public CoordinateConverter(BigDecimal[][] source, BigDecimal[][] destination) {
         this.converter = buildConverter(source, destination, false);
     }
 
-    public CoordinateConverter(Float[][] source, Float[][] destination, boolean checkCollinearity) {
+    public CoordinateConverter(BigDecimal[][] source, BigDecimal[][] destination, boolean checkCollinearity) {
         this.converter = buildConverter(source, destination, checkCollinearity);
     }
 
@@ -87,8 +84,8 @@ public class CoordinateConverter {
      * @param points Float[4][2] specifying the points for the 4 points, x & y.
      * @return boolean indicating collinearity
      */
-    static public boolean checkCollinearity(Float[][] points) {
-        return checkCollinearity(points, 0.00001f);
+    static public boolean checkCollinearity(BigDecimal[][] points) {
+        return checkCollinearity(points, new BigDecimal(.00001));
     }
 
     /**
@@ -106,7 +103,7 @@ public class CoordinateConverter {
      * @param eps Float indicating what passes for zero comparing the slopes.
      * @return boolean indicating collinearity
      */
-    static public boolean checkCollinearity(Float[][] points, Float eps) {
+    static public boolean checkCollinearity(BigDecimal[][] points, BigDecimal eps) {
         // 3 points (p1, p2, p3) are collinear if and only if
         //     abs( (p2.x-p1.x)*(p3.y-p1.y) -
         //          (p3.x-p1.x)*(p2.y-p1.y) ) <= eps
@@ -114,100 +111,133 @@ public class CoordinateConverter {
         // of lines sharing a point are the same (or very close), then
         // the lines are collinear.
 
-        Float x12 = points[1][0] - points[0][0];
-        Float y12 = points[1][1] - points[0][1];
-        Float x13 = points[2][0] - points[0][0];
-        Float y13 = points[2][1] - points[0][1];
-        Float x14 = points[3][0] - points[0][0];
-        Float y14 = points[3][1] - points[0][1];
-        Float x23 = points[2][0] - points[1][0];
-        Float y23 = points[2][1] - points[1][1];
-        Float x24 = points[3][0] - points[1][0];
-        Float y24 = points[3][1] - points[1][1];
+        BigDecimal x12 = points[1][0].subtract(points[0][0]);
+        BigDecimal y12 = points[1][1].subtract(points[0][1]);
+        BigDecimal x13 = points[2][0].subtract(points[0][0]);
+        BigDecimal y13 = points[2][1].subtract(points[0][1]);
+        BigDecimal x14 = points[3][0].subtract(points[0][0]);
+        BigDecimal y14 = points[3][1].subtract(points[0][1]);
+        BigDecimal x23 = points[2][0].subtract(points[1][0]);
+        BigDecimal y23 = points[2][1].subtract(points[1][1]);
+        BigDecimal x24 = points[3][0].subtract(points[1][0]);
+        BigDecimal y24 = points[3][1].subtract(points[1][1]);
         // Test each unique triplet.
         // 4 choose 3 = 4 triplets: 123, 124, 134, 234
-        return ((Math.abs(x12 * y13 - x13 * y12) < eps) ||
-                (Math.abs(x12 * y14 - x14 * y12) < eps) ||
-                (Math.abs(x13 * y14 - x14 * y13) < eps) ||
-                (Math.abs(x23 * y24 - x24 * y23) < eps));
+        return (x12.multiply(y13).subtract(x13.multiply(y12)).abs().compareTo(eps)) < 0 ||
+                (x12.multiply(y14).subtract(x14.multiply(y12)).abs().compareTo(eps) < 0) ||
+                (x13.multiply(y14).subtract(x14.multiply(y13)).abs().compareTo(eps) < 0 ||
+                (x23.multiply(y24).subtract(x24.multiply(y23))).abs().compareTo(eps) < 0);
     }
 
-    /**
-     * Construct the converter from the lists of floats.
-     *
-     * The converter lives for the life of the class.  It's a long-term object
-     * that is re-used using the <code>convert()</code> method.
-     *
-     * @param source 2D array of floats defining the source coordinate space
-     * @param destination 2D array of floats defining the destination coordinate space
-     * @param checkCollinearity boolean indicating if the constructor should check source & destination points for collinearity
-     * @return Mat holding the the constructed converter
-     */
-    private Mat buildConverter(Float[][] source, Float[][] destination, boolean checkCollinearity) {
-
-        if (checkCollinearity(source)) {
-            if (checkCollinearity) {
-                throw new IllegalArgumentException("Source points provided are collinear.");
-            } else {
-                log.warn("Source points are collinear (or close) which may cause problems in conversion");
-            }
-        } else if (checkCollinearity(destination)) {
-            if (checkCollinearity) {
-                throw new IllegalArgumentException("Destination points provided are collinear.");
-            } else {
-                log.warn("Destination points are collinear (or close) which may cause problems in conversion");
-            }
-        }
-        src = matixFromInput(source);
-        dst = matixFromInput(destination);
-
-        if (log.isTraceEnabled()) {
-            dumpMatrix(src, "src");
-            dumpMatrix(dst, "dst");
-        }
-
-        // DECOMP_LU (Gaussian elimination with the optimal pivot element chosen) is the default.
-        // We'll use the "long form" here should we need to add an option to override this in the future.
-
-        Mat cnvtr = Imgproc.getPerspectiveTransform(src, dst, Core.DECOMP_LU);
-        if (log.isTraceEnabled()) {
-            dumpMatrix(cnvtr, "Converter base");
-        }
-
-        return cnvtr;
+    BigDecimal[] adj(BigDecimal[] m) { // Compute the adjugate of m
+        return new BigDecimal[]{
+                m[4].multiply(m[8]).subtract(m[5].multiply(m[7])), m[2].multiply(m[7]).subtract(m[1].multiply(m[8])), m[1].multiply(m[5]).subtract(m[2].multiply(m[4])),
+                m[5].multiply(m[6]).subtract(m[3].multiply(m[8])), m[0].multiply(m[8]).subtract(m[2].multiply(m[6])), m[2].multiply(m[3]).subtract(m[0].multiply(m[5])),
+                m[3].multiply(m[7]).subtract(m[4].multiply(m[6])), m[1].multiply(m[6]).subtract(m[0].multiply(m[7])), m[0].multiply(m[4]).subtract(m[1].multiply(m[3])),
+        };
     }
 
-    private Mat matixFromInput(Float[][] flts) {
-        Mat result = new Mat(new Size(1, 4), CV_32FC2);
-        float maxX = 0, minX = 0, maxY = 0, minY = 0;
-
-        for (int i = 0; i < flts.length; i++) {
-            if (i == 0) {
-                maxX = minX = flts[i][0];
-                maxY = minY = flts[i][1];
-
-            } else {
-                if (maxX < flts[i][0]) {
-                    maxX = flts[i][0];
+    BigDecimal[] multmm(BigDecimal[] a, BigDecimal[] b) { // multiply two matrices
+        BigDecimal[] c = new BigDecimal[9];
+        for (int i = 0; i < 3; i++) {
+            for (int j = 0; j < 3; j++) {
+                BigDecimal cij = new BigDecimal(0);
+                for (int k = 0; k < 3; k++) {
+                    cij = cij.add(a[3 * i + k].multiply(b[3 * k + j]));
                 }
-                if (minX > flts[i][0]) {
-                    minX = flts[i][0];
-                }
-                if (maxY < flts[i][1]) {
-                    maxY = flts[i][1];
-                }
-                if (minY > flts[i][1]) {
-                    minY = flts[i][1];
-                }
-
+                c[3 * i + j] = cij;
             }
-            result.put(i, 0, flts[i][0], flts[i][1]);
         }
-
-        dumpMatrix(result, "input Matrix");
-        log.debug("X difference: {}, y difference: {}", maxX - minX, maxY - minY);
-        return result;
+        return c;
     }
+
+    BigDecimal[] multmv(BigDecimal[] m, BigDecimal[] v) { // multiply matrix and vector
+        return new BigDecimal[]{
+                m[0].multiply(v[0]).add(m[1].multiply(v[1])).add(m[2].multiply(v[2])),
+                m[3].multiply(v[0]).add(m[4].multiply(v[1])).add(m[5].multiply(v[2])),
+                m[6].multiply(v[0]).add(m[7].multiply(v[1])).add(m[8].multiply(v[2])),
+        };
+    }
+
+    BigDecimal[] basisToPoints(BigDecimal x1, BigDecimal y1, BigDecimal x2, BigDecimal y2, BigDecimal x3, BigDecimal y3, BigDecimal x4, BigDecimal y4) {
+        BigDecimal[] m = new BigDecimal[]{
+                x1, x2, x3,
+                y1, y2, y3,
+                BigDecimal.ONE, BigDecimal.ONE, BigDecimal.ONE
+        };
+        BigDecimal[] v = multmv(adj(m),new BigDecimal[]{x4, y4, BigDecimal.ONE});
+        return multmm(m,new BigDecimal[]{
+                v[0], BigDecimal.ZERO, BigDecimal.ZERO,
+                BigDecimal.ZERO, v[1], BigDecimal.ZERO,
+                BigDecimal.ZERO, BigDecimal.ZERO, v[2]
+        });
+    }
+
+    BigDecimal[] general2DProjection(
+            BigDecimal x1s, BigDecimal y1s, BigDecimal x1d, BigDecimal y1d,
+            BigDecimal x2s, BigDecimal y2s, BigDecimal x2d, BigDecimal y2d,
+            BigDecimal x3s, BigDecimal y3s, BigDecimal x3d, BigDecimal y3d,
+            BigDecimal x4s, BigDecimal y4s, BigDecimal x4d, BigDecimal y4d
+    ) {
+        BigDecimal[] s = basisToPoints(x1s, y1s, x2s, y2s, x3s, y3s, x4s, y4s);
+        BigDecimal[] d = basisToPoints(x1d, y1d, x2d, y2d, x3d, y3d, x4d, y4d);
+        return multmm(d, adj(s));
+    }
+
+    BigDecimal[] project(BigDecimal[] m, BigDecimal x, BigDecimal y) {
+        BigDecimal[] v = multmv(m, new BigDecimal[]{x, y, BigDecimal.ONE});
+        return new BigDecimal[]{v[0].divide(v[2], BigDecimal.ROUND_CEILING), v[1].divide(v[2], BigDecimal.ROUND_CEILING)};
+    }
+ /**
+  * Construct the converter from the lists of floats.
+  *
+  * The converter lives for the life of the class.  It's a long-term object
+  * that is re-used using the <code>convert()</code> method.
+  *
+  * @param source 2D array of floats defining the source coordinate space
+  * @param destination 2D array of floats defining the destination coordinate space
+  * @param checkCollinearity boolean indicating if the constructor should check source & destination points for collinearity
+  * @return Mat holding the the constructed converter
+  */
+    private BigDecimal[] buildConverter(BigDecimal[][] source, BigDecimal [][] destination, boolean checkCollinearity) {
+        return general2DProjection(
+                source[0][0], source[0][1], destination[0][0], destination[0][1],
+                source[1][0], source[1][1], destination[1][0], destination[1][1],
+                source[2][0], source[2][1], destination[2][0], destination[2][1],
+                source[3][0], source[3][1], destination[3][0], destination[3][1]);
+    }
+
+//    private Mat matixFromInput(Float[][] flts) {
+//        Mat result = new Mat(new Size(1, 4), CV_32FC2);
+//        float maxX = 0, minX = 0, maxY = 0, minY = 0;
+//
+//        for (int i = 0; i < flts.length; i++) {
+//            if (i == 0) {
+//                maxX = minX = flts[i][0];
+//                maxY = minY = flts[i][1];
+//
+//            } else {
+//                if (maxX < flts[i][0]) {
+//                    maxX = flts[i][0];
+//                }
+//                if (minX > flts[i][0]) {
+//                    minX = flts[i][0];
+//                }
+//                if (maxY < flts[i][1]) {
+//                    maxY = flts[i][1];
+//                }
+//                if (minY > flts[i][1]) {
+//                    minY = flts[i][1];
+//                }
+//
+//            }
+//            result.put(i, 0, flts[i][0], flts[i][1]);
+//        }
+//
+//        dumpMatrix(result, "input Matrix");
+//        log.debug("X difference: {}, y difference: {}", maxX - minX, maxY - minY);
+//        return result;
+//    }
 
     /**
      * Convert a 2D coordinate according this converter's specification
@@ -220,76 +250,24 @@ public class CoordinateConverter {
      * @param srcCoordsArray Array[2] of Floats that represent the coordinate to be converted.
      * @return Array[2] of Floats representing the result of the conversion.
      */
-    public Float[] convert(Float[] srcCoordsArray) {
-        // (Object) cast to force log.debug() to understand that it's not a varargs thing....
-        log.debug("convert({}) called...", (Object) srcCoordsArray);
-
-        // Construct a matrix (OpenCV Mat) representation of our input coordinates for conversion.
-        // Note that we construct the transpose of what would normally be used for the coordinates
-        // since the conversion operation involves (matrix) multiplication of our converter
-        // with the transpose of the coordinates.  Since that's all we ever do with them & things are small,
-        // we'll just create the transpose manually & use that directly.
-        //
-        // Calling this out so that readers understand why the matrix is created this way.
-
-        Mat coords = new Mat(3,1, CV_64FC1);
-        coords.put(0, 0, srcCoordsArray[0]);
-        coords.put(1, 0, srcCoordsArray[1]);
-        coords.put(2, 0, 1d);   // Fill in an identity value to ease matrix math
-        if (log.isTraceEnabled()) {
-            dumpMatrix(coords, "coords");   // Dump out for debug purposes
-        }
-
-        // Here, it's a bit tricky.
-        // Here, we're going to perform a matrix multiply of the converter with
-        // the (transpose of) our coordinates.  As noted above, the coordinates (coords) were created
-        // transposed, so there's no extra work here.  Also, as noted in the class constructor, the types
-        // in use here need to match, so we use a the coordinateConverter that's been converted to the types
-        // used for our coordinates.
-        //
-        // To perform this multiply, we'll use the Generalized Matrix Multiplication (gemm) method.
-        // Core.gemm(src1, src2, alpha, src3, beta, dst) performs the operation as:
-        //     dst = alpha*src1.t()*src2 + beta*src3.t()
-        // in python speak, which is to say that the result is the product of the (alphas * src1.transpose()) times
-        // src2, to which is added the product of beta * src3.transpose();
-        //
-        // In our case, we don't need any weighted version, so the alpha parameter (#3) is 1.  Moreover, we
-        // need nothing added to the results, so beta is zero, and the src3 parameter is an empty matrix.
-        // Despite beta being zero, we must pass in an empty src3 matrix (parameter #4) that would've been
-        // used in association with the beta parameter to adjust the result.
-        //
-        // The result of this operation appear in our resultCoords matrix, from which we extract the converter
-        // coordinates.
-
-        // Construct a Mat to hold the results of the transform
-        Mat resultCoords = new Mat();
-        // Do the math...
-        Core.gemm(converter, coords, 1, Mat.zeros(converter.size(), converter.type()), 0, resultCoords);
-
-        if (log.isTraceEnabled()) {
-            dumpMatrix(resultCoords, "resultCoords");
-        }
-
-        // Convert result matrix into a simple array of coordinates.
-        Float[] result = new Float[] {(float) resultCoords.get(0,0)[0], (float) resultCoords.get(1,0)[0]};
-        log.debug("convert({}) --> {}", srcCoordsArray, result);
-
-        return result;
-    }
-
-    /**
-     * Dump an OpenCV::Mat value for debug purposes
-     * @param m Mat to be dumped
-     * @param name String name of Mat (or other descriptive phrase of value to developer)
-     */
-    private void dumpMatrix(Mat m, String name) {
-        log.debug(name + " description: {} (type: {}, size: {})", m.toString(), m.type(), m.size());
-        log.debug("Contents of {}", name);
-        for (int row = 0; row < m.rows(); row++) {
-            for (int col = 0; col < m.cols(); col++) {
-                log.debug("\trow {}, col {} :: {}", row, col, m.get(row, col));
-            }
-        }
+    public BigDecimal[] convert(BigDecimal [] srcCoordsArray) {
+        return project(converter, srcCoordsArray[0], srcCoordsArray[1]);
 
     }
+
+//    /**
+//     * Dump an OpenCV::Mat value for debug purposes
+//     * @param m Mat to be dumped
+//     * @param name String name of Mat (or other descriptive phrase of value to developer)
+//     */
+//    private void dumpMatrix(Mat m, String name) {
+//        log.debug(name + " description: {} (type: {}, size: {})", m.toString(), m.type(), m.size());
+//        log.debug("Contents of {}", name);
+//        for (int row = 0; row < m.rows(); row++) {
+//            for (int col = 0; col < m.cols(); col++) {
+//                log.debug("\trow {}, col {} :: {}", row, col, m.get(row, col));
+//            }
+//        }
+//
+//    }
 }
