+++ objectRecognitionSource/src/test/java/io/vantiq/extsrc/objectRecognition/imageRetriever/TestCoordConverter.java
@@ -9,24 +9,20 @@
 package io.vantiq.extsrc.objectRecognition.imageRetriever;
 
 import lombok.extern.slf4j.Slf4j;
-import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
-import org.opencv.core.Core;
+
+import java.math.BigDecimal;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertNotNull;
 
 @Slf4j
 public class TestCoordConverter {
 
-    static final Float ACCEPTABLE_DELTA = 0.8f;
-
-    @BeforeClass
-    public static void loadEmUp() {
-        System.loadLibrary(Core.NATIVE_LIBRARY_NAME);
-    }
+    static final BigDecimal ACCEPTABLE_DELTA = new BigDecimal(0.00000001d);
+    
 
     // FIXME:  Would be good to get some "real" GPS conversions in unit tests since that's really
     // FIXME: the primary goal here.
@@ -35,12 +31,89 @@ public class TestCoordConverter {
     public void testIdentity() {
 
         // Test the simplest of conversions
-        Float[][] srcPts = new Float[][]{{1.0f, 1f}, {2f, 1f}, {3f, 3f}, {4f, 3f}};
-        Float[][] dstPts = new Float[][]{{1.0f, 1f}, {2f, 1f}, {3f, 3f}, {4f, 3f}};
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
 
-        Float[][] tstPts = generateRandomPoints(100);
-        Float[][] tstResPts = tstPts.clone();
+        BigDecimal[][] tstPts = generateRandomPoints(100);
+        BigDecimal[][] tstResPts = tstPts.clone();
 
+        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
+    }
+
+    @Test
+    public void testFlipY() {
+
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(-1d)},
+                {new BigDecimal(2d), new BigDecimal(-1d)},
+                {new BigDecimal(3d), new BigDecimal(-3d)},
+                {new BigDecimal(4d), new BigDecimal(-3d)}};
+
+        BigDecimal[][] tstPts = generateRandomPoints(100);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
+        for (int i = 0; i < tstPts.length; i++) {
+            tstResPts[i][0] = tstPts[i][0];
+            tstResPts[i][1] = BigDecimal.ZERO.subtract(tstPts[i][1]);
+        }
+        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
+    }
+
+    @Test
+    public void testFlipX() {
+
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(-1.0d), new BigDecimal(1d)},
+                {new BigDecimal(-2d), new BigDecimal(1d)},
+                {new BigDecimal(-3d), new BigDecimal(3d)},
+                {new BigDecimal(-4d), new BigDecimal(3d)}};
+
+        BigDecimal[][] tstPts = generateRandomPoints(100);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
+        for (int i = 0; i < tstPts.length; i++) {
+            tstResPts[i][0] = BigDecimal.ZERO.subtract(tstPts[i][0]);
+            tstResPts[i][1] = tstPts[i][1];
+        }
+        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
+    }
+
+    @Test
+    public void testInvertImage() {
+
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(-1.0d), new BigDecimal(-1d)},
+                {new BigDecimal(-2d), new BigDecimal(-1d)},
+                {new BigDecimal(-3d), new BigDecimal(-3d)},
+                {new BigDecimal(-4d), new BigDecimal(-3d)}};
+
+        BigDecimal[][] tstPts = generateRandomPoints(100);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
+        for (int i = 0; i < tstPts.length; i++) {
+            tstResPts[i][0] = BigDecimal.ZERO.subtract(tstPts[i][0]);
+            tstResPts[i][1] = BigDecimal.ZERO.subtract(tstPts[i][1]);
+        }
         performConverterTest(srcPts, dstPts, tstPts, tstResPts);
     }
 
@@ -48,22 +121,23 @@ public class TestCoordConverter {
     public void test50Plus() {
         // Test change of base -- still a linear conversion (add 50)
 
-        Float[][] srcPts = new Float[][]{
-                {3f, 3f},
-                {1.0f, 1f},
-                {2f, 1f},
-                {4f, 3f}};
-        Float[][] dstPts = new Float[][]{
-                {53f, 53f},
-                {51.0f, 51f},
-                {52f, 51f},
-                {54f, 53f}};
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(3), new BigDecimal(3)},
+                {new BigDecimal(1.0d), new BigDecimal(1)},
+                {new BigDecimal(2), new BigDecimal(1)},
+                {new BigDecimal(4), new BigDecimal(3)}
+        };
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(53d), new BigDecimal(53d)},
+                {new BigDecimal(51.0d), new BigDecimal(51d)},
+                {new BigDecimal(52d), new BigDecimal(51d)},
+                {new BigDecimal(54d), new BigDecimal(53d)}};
 
-        Float[][] tstPts = generateRandomPoints(100);
-        Float[][] tstResPts = new Float[tstPts.length][2];
+        BigDecimal[][] tstPts = generateRandomPoints(100);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
         for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0] + 50;
-            tstResPts[i][1] = tstPts[i][1] + 50;
+            tstResPts[i][0] = tstPts[i][0].add(new BigDecimal(50));
+            tstResPts[i][1] = tstPts[i][1].add(new BigDecimal(50));
         }
 
         performConverterTest(srcPts, dstPts, tstPts, tstResPts);
@@ -73,14 +147,26 @@ public class TestCoordConverter {
     public void testScaleChange() {
         // Test a conversion that changes the scale of both the x & y axes.
 
-        Float[][] srcPts = new Float[][]{{1.0f, 1f}, {2f, 1f}, {3f, 3f}, {4f, 3f}};
-        Float[][] dstPts = new Float[][]{{2.0f, .002f}, {4f, .002f}, {6f, .006f}, {8f, .006f}};
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}
+        };
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(2.0d), new BigDecimal(.002d)},
+                {new BigDecimal(4d), new BigDecimal(.002d)},
+                {new BigDecimal(6d), new BigDecimal(.006d)},
+                {new BigDecimal(8d), new BigDecimal(.006d)}
+        };
 
-        Float[][] tstPts = generateRandomPoints(200);
-        Float[][] tstResPts = new Float[tstPts.length][2];
+        BigDecimal[][] tstPts = generateRandomPoints(200);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
         for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0] * 2;
-            tstResPts[i][1] = (tstPts[i][1] * 2) / 1000;
+            tstResPts[i][0] = tstPts[i][0].multiply(new BigDecimal(2));
+            tstResPts[i][1] = tstPts[i][1].multiply(new BigDecimal(2))
+                    .divide(new BigDecimal(1000),
+                            BigDecimal.ROUND_HALF_UP);
         }
         performConverterTest(srcPts, dstPts, tstPts, tstResPts);
 
@@ -90,34 +176,42 @@ public class TestCoordConverter {
     public void testWarpChange() {
         // Test a conversion that changes the scale of only the x axis
 
-        Float[][] srcPts = new Float[][]{
-                {3f, 3f},
-                {2f, 1f},
-                {1.0f, 1f},
-                {4f, 3f}};
-        Float[][] dstPts = new Float[][]{
-                {6f, 3f},
-                {4f, 1f},
-                {2.0f, 1f},
-                {8f, 3f}};
+        BigDecimal[][] srcPts = new BigDecimal[][]{
+                {new BigDecimal(3), new BigDecimal(3)},
+                {new BigDecimal(2), new BigDecimal(1)},
+                {new BigDecimal(1.0), new BigDecimal(1)},
+                {new BigDecimal(4), new BigDecimal(3)}};
+        BigDecimal[][] dstPts = new BigDecimal[][]{
+                {new BigDecimal(6d), new BigDecimal(3)},
+                {new BigDecimal(4d), new BigDecimal(1d)},
+                {new BigDecimal(2.0), new BigDecimal(1d)},
+                {new BigDecimal(8d), new BigDecimal(3d)}};
 
-        Float[][] tstPts = generateRandomPoints(200);
-        Float[][] tstResPts = new Float[tstPts.length][2];
+        BigDecimal[][] tstPts = generateRandomPoints(200);
+        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
         for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0] * 2;
+            tstResPts[i][0] = tstPts[i][0].multiply(new BigDecimal(2));
             tstResPts[i][1] = tstPts[i][1];
         }
         performConverterTest(srcPts, dstPts, tstPts, tstResPts);
 
-        srcPts = new Float[][]{{1.0f, 1f}, {2f, 1f}, {3f, 3f}, {4f, 3f}};
-        dstPts = new Float[][]{{1.0f, 10f}, {2f, 10f}, {3f, 30f}, {4f, 30f}};
+        srcPts = new BigDecimal[][]{
+                {new BigDecimal(1.0), new BigDecimal(1d)},
+                {new BigDecimal(2d), new BigDecimal(1d)},
+                {new BigDecimal(3d), new BigDecimal(3d)},
+                {new BigDecimal(4d), new BigDecimal(3d)}};
+        dstPts = new BigDecimal[][]{
+                {new BigDecimal(1.0d), new BigDecimal(10d)},
+                {new BigDecimal(2d), new BigDecimal(10d)},
+                {new BigDecimal(3d), new BigDecimal(30d)},
+                {new BigDecimal(4d), new BigDecimal(30d)}};
 
         // Now, let's do a warp that changes only the y axis
         tstPts = generateRandomPoints(200);
-        tstResPts = new Float[tstPts.length][2];
+        tstResPts = new BigDecimal[tstPts.length][2];
         for (int i = 0; i < tstPts.length; i++) {
             tstResPts[i][0] = tstPts[i][0];
-            tstResPts[i][1] = tstPts[i][1] * 10;
+            tstResPts[i][1] = tstPts[i][1].multiply(new BigDecimal(10));
         }
         performConverterTest(srcPts, dstPts, tstPts, tstResPts);
     }
@@ -125,23 +219,18 @@ public class TestCoordConverter {
     @Test
     public void testGpsConv() {
         // image location --> lat/long
-        Float[][] gpsSrcPts = new Float[][]{
-                {1205f, 698f}, // Corner of dashed line bottom right
-                {87f, 328f}, // Corner of white rumble strip top left
-                {1200f, 278f}, // Third lamppost top right
-                {36f, 583f}, // Corner of rectangular road marking bottom left
+        BigDecimal[][] gpsSrcPts = new BigDecimal[][]{
+                {new BigDecimal(1205), new BigDecimal(698d)}, // Corner of dashed line bottom right
+                {new BigDecimal(87d), new BigDecimal(328d)}, // Corner of white rumble strip top left
+                {new BigDecimal(1200d), new BigDecimal(278d)}, // Third lamppost top right
+                {new BigDecimal(36d), new BigDecimal(583d)}, // Corner of rectangular road marking bottom left
         };
         // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        Float[][] gpsDstPts = new Float[][]{
-//                {6.603560f, 52.036730f * 1000}, // Corner of dashed line bottom right
-//                {6.603227f, 52.036181f * 1000}, // Corner of white rumble strip top left
-//                {6.602018f, 52.036769f * 1000}, // Third lamppost top right
-//                {6.603638f, 52.036558f * 1000}, // Corner of rectangular road marking bottom left
-                {.003560f, .036730f}, // Corner of dashed line bottom right
-                {.003227f, .036181f}, // Corner of white rumble strip top left
-                {.002018f, .036769f}, // Third lamppost top right
-                {.003638f, .036558f}, // Corner of rectangular road marking bottom left
-
+        BigDecimal[][] gpsDstPts = new BigDecimal[][]{
+                {new BigDecimal(6.603560), new BigDecimal(52.036730d)}, // Corner of dashed line bottom right
+                {new BigDecimal(6.603227), new BigDecimal(52.036181d)}, // Corner of white rumble strip top left
+                {new BigDecimal(6.602018), new BigDecimal(52.036769d)}, // Third lamppost top right
+                {new BigDecimal(6.603638), new BigDecimal(52.036558d)}, // Corner of rectangular road marking bottom left
         };
 
         performConverterTest(gpsSrcPts, gpsDstPts);
@@ -149,24 +238,20 @@ public class TestCoordConverter {
 
     @Test
     public void testAlleyCam1Collinearity() {
-        Float[][] alleyCam1SrcPts = new Float[][]{
-                {65.f, 10.f},
-                {847.f, 8.f},
-                {850.f, 694.f},
-                {63.f, 696.f},
+        BigDecimal[][] alleyCam1SrcPts = new BigDecimal[][]{
+                {new BigDecimal(65.d), new BigDecimal(10.d)},
+                {new BigDecimal(847.d), new BigDecimal(8.d)},
+                {new BigDecimal(850.d), new BigDecimal(694.d)},
+                {new BigDecimal(63.d), new BigDecimal(696.d)},
         };
         // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        Float[][] alleyCam1DstPts = new Float[][]{
-                {-83.74801820260836f, 42.27796103748212f},
-                {-83.74819690478398f, 42.27796396568019f},
-                {-83.74813303468181f, 42.78190293489615f},
-                {-83.74800177407815f, 42.78187357400434f},
-//                {-.74801820260836f, .27796103748212f},
-//                {-.74819690478398f, .27796396568019f},
-//                {-.74813303468181f, .78190293489615f},
-//                {-.74800177407815f, .78187357400434f},
+        BigDecimal[][] alleyCam1DstPts = new BigDecimal[][]{
+                {new BigDecimal(-83.74801820260836), new BigDecimal(42.27796103748212)},
+                {new BigDecimal(-83.74819690478398), new BigDecimal(42.27796396568019)},
+                {new BigDecimal(-83.74813303468181), new BigDecimal(42.78190293489615)},
+                {new BigDecimal(-83.74800177407815), new BigDecimal(42.78187357400434)},
         };
-        assertFalse("Destination Points Collinear", CoordinateConverter.checkCollinearity(alleyCam1DstPts, 0.00001f));
+        assertFalse("Destination Points Collinear", CoordinateConverter.checkCollinearity(alleyCam1DstPts, new BigDecimal(0.00001d)));
     }
 
     @Test
@@ -174,35 +259,31 @@ public class TestCoordConverter {
     public void testAlleyCam1() {
         // image location --> lat/long
         // Coords in order topLeft, topRight, bottomRight, bottomLeft (CW from topLeft)
-        Float[][] alleyCam1SrcPts = new Float[][]{
-                {65.f, 10.f},
-                {847.f, 8.f},
-                {850.f, 694.f},
-                {63.f, 696.f},
+        BigDecimal[][] alleyCam1SrcPts = new BigDecimal[][]{
+                {new BigDecimal(65.d), new BigDecimal(10.d)},
+                {new BigDecimal(847.d), new BigDecimal(8.d)},
+                {new BigDecimal(850.d), new BigDecimal(694.d)},
+                {new BigDecimal(63.d), new BigDecimal(696.d)},
         };
             // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        Float[][] alleyCam1DstPts = new Float[][]{
-//                {-83.74801820260836f, 42.27796103748212f},
-//                {-83.74819690478398f, 42.27796396568019f},
-//                {-83.74813303468181f, 42.78190293489615f},
-//                {-83.74800177407815f, 42.78187357400434f},
-                {-.74801820260836f, .27796103748212f},
-                {-.74819690478398f, .27796396568019f},
-                {-.74813303468181f, .78190293489615f},
-                {-.74800177407815f, .78187357400434f},
+        BigDecimal[][] alleyCam1DstPts = new BigDecimal[][]{
+                {new BigDecimal(-83.74801820260836), new BigDecimal(42.27796103748212)},
+                {new BigDecimal(-83.74819690478398), new BigDecimal(42.27796396568019)},
+                {new BigDecimal(-83.74813303468181), new BigDecimal(42.78190293489615)},
+                {new BigDecimal(-83.74800177407815), new BigDecimal(42.78187357400434)},
         };
 
         assertFalse("Source Points Collinear", CoordinateConverter.checkCollinearity(alleyCam1SrcPts));
-        assertFalse("Destination Points Collinear", CoordinateConverter.checkCollinearity(alleyCam1DstPts, 0.00001f));
+        assertFalse("Destination Points Collinear", CoordinateConverter.checkCollinearity(alleyCam1DstPts, new BigDecimal(0.00001d)));
 
         performConverterTest(alleyCam1SrcPts, alleyCam1DstPts);
     }
 
-    protected void performConverterTest(Float[][] srcPts, Float[][] dstPts) {
+    protected void performConverterTest(BigDecimal[][] srcPts, BigDecimal[][] dstPts) {
         performConverterTest(srcPts, dstPts, null, null);
     }
 
-    protected void performConverterTest(Float[][] srcPts, Float[][] dstPts, Float[][] tstSrc, Float[][] tstRes) {
+    protected void performConverterTest(BigDecimal[][] srcPts, BigDecimal[][] dstPts, BigDecimal[][] tstSrc, BigDecimal[][] tstRes) {
 
         // First, create a converter calibrated by the src & dst points passed in.
         CoordinateConverter converter = new CoordinateConverter(srcPts, dstPts);
@@ -211,15 +292,17 @@ public class TestCoordConverter {
 
         for (int iter = 0; iter < 2; iter++) {
             for (int i = 0; i < srcPts.length; i++) {
-                Float[] src = srcPts[i];
-                Float[] dst = dstPts[i];
-                Float[] res = converter.convert(src);
+                BigDecimal[] src = srcPts[i];
+                BigDecimal[] dst = dstPts[i];
+                BigDecimal[] res = converter.convert(src);
                 if (iter ==1) {
-                    assertEquals("X coord mismatch", dst[0], res[0], ACCEPTABLE_DELTA);
-                    assertEquals("Y coord mismatch", dst[1], res[1], ACCEPTABLE_DELTA);
+                    assertTrue("X coord mismatch (" + i + "): expected: " + dst[0] + ", found: " + res[0],
+                            dst[0].compareTo(res[0]) == 0);
+                    assertTrue("Y coord mismatch (" + i + "): expected: " + dst[1] + ", found: " + res[1],
+                            dst[1].compareTo(res[1]) == 0);
                 } else {
-                    log.debug("Iter: {}:{}, X: dst: {}, res:{}, diff: {}", iter, i, dst[0], res[0], dst[0] - res[0]);
-                    log.debug("Iter: {}:{}, Y: dst: {}, res:{}, diff: {}", iter, i, dst[1], res[1], dst[1] - res[1]);
+                    log.debug("Iter: {}:{}, X: dst: {}, res:{}, diff: {}", iter, i, dst[0], res[0], dst[0].subtract(res[0]));
+                    log.debug("Iter: {}:{}, Y: dst: {}, res:{}, diff: {}", iter, i, dst[1], res[1], dst[1].subtract(res[1]));
                 }
             }
         }
@@ -232,26 +315,34 @@ public class TestCoordConverter {
                     tstSrc.length, tstRes.length);
 
             for (int ti = 0; ti < tstSrc.length; ti++) {
-                Float[] src = tstSrc[ti];
-                Float[] dst = tstRes[ti];
+                BigDecimal[] src = tstSrc[ti];
+                BigDecimal[] dst = tstRes[ti];
                 log.debug("Checking conversion of ({}, {})", src[0], src[1]);
-                Float[] res = converter.convert(src);
-                assertEquals("X coord mismatch in test points index: " + ti, dst[0], res[0], ACCEPTABLE_DELTA);
-                assertEquals("Y coord mismatch in test points index: " + ti, dst[1], res[1], ACCEPTABLE_DELTA);
+                BigDecimal[] res = converter.convert(src);
+
+                // Crazy test generation sometimes comes up with odd values.  Scale numbers to reasonable
+                for (int i = 0; i < 2; i++) {
+                    res[i] = res[i].setScale(dst[i].scale(), BigDecimal.ROUND_HALF_UP);
+                }
+                assertTrue("X coord mismatch in test points index: " + ti + ": expected: " + dst[0] + ", found: " + res[0],
+                        dst[0].compareTo(res[0]) == 0);
+                assertTrue("Y coord mismatch in test points index: " + ti + ": expected: " + dst[1] + ", found: " + res[1],
+                        dst[1].compareTo(res[1]) == 0);
             }
         }
     }
 
-    public static Float[][] generateRandomPoints(int pointCount) {
-        Float[][] points = new Float[pointCount][2];
+    public static BigDecimal[][] generateRandomPoints(int pointCount) {
+        BigDecimal[][] points = new BigDecimal[pointCount][2];
         for (int i = 0; i < pointCount; i++) {
             // Generate some random points all over the place to test against...
-            Float factor1 = (i % 5 == 0 ? 1f : 10f * (i % 7 == 0 ? pointCount : 1));
-            Float x = (float) (Math.random() * factor1);
+            BigDecimal factor1 = new BigDecimal(i % 5 == 0 ? 1d : 10d * (i % 7 == 0 ? pointCount : 1));
+            BigDecimal x = factor1.multiply(new BigDecimal(Math.random()));
 
-            Float factor2 = (i % 2 == 0 ? 1f : 130f * (i % 3 == 0 ? pointCount : 1));
-            Float y = (float) (Math.random() * factor1);
-            points[i] = new Float[]{x, y};
+            BigDecimal factor2 = new BigDecimal(i % 2 == 0 ? 1d : 130d * (i % 3 == 0 ? pointCount : 1));
+            // Limit craziness of test comparison for randomly generated values...
+            BigDecimal y = factor2.multiply(new BigDecimal(Math.random())).setScale(12, BigDecimal.ROUND_HALF_UP);
+            points[i] = new BigDecimal[]{x, y};
         }
         return points;
     }
