--- objectRecognitionSource/src/test/java/io/vantiq/extsrc/objectRecognition/imageRetriever/TestCoordConverter2.java
@@ -1,364 +1,0 @@
-/*
- * Copyright (c) 2019 Vantiq, Inc.
- *
- * All rights reserved.
- *
- * SPDX: MIT
- */
-
-package io.vantiq.extsrc.objectRecognition.imageRetriever;
-
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
-
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.math.MathContext;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.assertNotNull;
-
-@Slf4j
-public class TestCoordConverter2 {
-
-    static final BigDecimal ACCEPTABLE_DELTA = new BigDecimal(0.00000001d);
-    
-
-    // FIXME:  Would be good to get some "real" GPS conversions in unit tests since that's really
-    // FIXME: the primary goal here.
-
-    @Test
-    public void testIdentity() {
-
-        // Test the simplest of conversions
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(100);
-        BigDecimal[][] tstResPts = tstPts.clone();
-
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void testFlipY() {
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(-1d)},
-                {new BigDecimal(2d), new BigDecimal(-1d)},
-                {new BigDecimal(3d), new BigDecimal(-3d)},
-                {new BigDecimal(4d), new BigDecimal(-3d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(100);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0];
-            tstResPts[i][1] = BigDecimal.ZERO.subtract(tstPts[i][1]);
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void testFlipX() {
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(-1.0d), new BigDecimal(1d)},
-                {new BigDecimal(-2d), new BigDecimal(1d)},
-                {new BigDecimal(-3d), new BigDecimal(3d)},
-                {new BigDecimal(-4d), new BigDecimal(3d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(100);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = BigDecimal.ZERO.subtract(tstPts[i][0]);
-            tstResPts[i][1] = tstPts[i][1];
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void testInvertImage() {
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(-1.0d), new BigDecimal(-1d)},
-                {new BigDecimal(-2d), new BigDecimal(-1d)},
-                {new BigDecimal(-3d), new BigDecimal(-3d)},
-                {new BigDecimal(-4d), new BigDecimal(-3d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(100);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = BigDecimal.ZERO.subtract(tstPts[i][0]);
-            tstResPts[i][1] = BigDecimal.ZERO.subtract(tstPts[i][1]);
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void test50Plus() {
-        // Test change of base -- still a linear conversion (add 50)
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(3), new BigDecimal(3)},
-                {new BigDecimal(1.0d), new BigDecimal(1)},
-                {new BigDecimal(2), new BigDecimal(1)},
-                {new BigDecimal(4), new BigDecimal(3)}
-        };
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(53d), new BigDecimal(53d)},
-                {new BigDecimal(51.0d), new BigDecimal(51d)},
-                {new BigDecimal(52d), new BigDecimal(51d)},
-                {new BigDecimal(54d), new BigDecimal(53d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(100);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0].add(new BigDecimal(50));
-            tstResPts[i][1] = tstPts[i][1].add(new BigDecimal(50));
-        }
-
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void testScaleChange() {
-        // Test a conversion that changes the scale of both the x & y axes.
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}
-        };
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(2.0d), new BigDecimal(.002d)},
-                {new BigDecimal(4d), new BigDecimal(.002d)},
-                {new BigDecimal(6d), new BigDecimal(.006d)},
-                {new BigDecimal(8d), new BigDecimal(.006d)}
-        };
-
-        BigDecimal[][] tstPts = generateRandomPoints(200);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0].multiply(new BigDecimal(2));
-            tstResPts[i][1] = tstPts[i][1].multiply(new BigDecimal(2))
-                    .divide(new BigDecimal(1000),
-                            BigDecimal.ROUND_HALF_UP);
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-
-    }
-
-    @Test
-    public void testWarpChange() {
-        // Test a conversion that changes the scale of only the x axis
-
-        BigDecimal[][] srcPts = new BigDecimal[][]{
-                {new BigDecimal(3), new BigDecimal(3)},
-                {new BigDecimal(2), new BigDecimal(1)},
-                {new BigDecimal(1.0), new BigDecimal(1)},
-                {new BigDecimal(4), new BigDecimal(3)}};
-        BigDecimal[][] dstPts = new BigDecimal[][]{
-                {new BigDecimal(6d), new BigDecimal(3)},
-                {new BigDecimal(4d), new BigDecimal(1d)},
-                {new BigDecimal(2.0), new BigDecimal(1d)},
-                {new BigDecimal(8d), new BigDecimal(3d)}};
-
-        BigDecimal[][] tstPts = generateRandomPoints(200);
-        BigDecimal[][] tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0].multiply(new BigDecimal(2));
-            tstResPts[i][1] = tstPts[i][1];
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-
-        srcPts = new BigDecimal[][]{
-                {new BigDecimal(1.0), new BigDecimal(1d)},
-                {new BigDecimal(2d), new BigDecimal(1d)},
-                {new BigDecimal(3d), new BigDecimal(3d)},
-                {new BigDecimal(4d), new BigDecimal(3d)}};
-        dstPts = new BigDecimal[][]{
-                {new BigDecimal(1.0d), new BigDecimal(10d)},
-                {new BigDecimal(2d), new BigDecimal(10d)},
-                {new BigDecimal(3d), new BigDecimal(30d)},
-                {new BigDecimal(4d), new BigDecimal(30d)}};
-
-        // Now, let's do a warp that changes only the y axis
-        tstPts = generateRandomPoints(200);
-        tstResPts = new BigDecimal[tstPts.length][2];
-        for (int i = 0; i < tstPts.length; i++) {
-            tstResPts[i][0] = tstPts[i][0];
-            tstResPts[i][1] = tstPts[i][1].multiply(new BigDecimal(10));
-        }
-        performConverterTest(srcPts, dstPts, tstPts, tstResPts);
-    }
-
-    @Test
-    public void testGpsConv() {
-        // image location --> lat/long
-        BigDecimal[][] gpsSrcPts = new BigDecimal[][]{
-                {new BigDecimal(1205), new BigDecimal(698d)}, // Corner of dashed line bottom right
-                {new BigDecimal(87d), new BigDecimal(328d)}, // Corner of white rumble strip top left
-                {new BigDecimal(1200d), new BigDecimal(278d)}, // Third lamppost top right
-                {new BigDecimal(36d), new BigDecimal(583d)}, // Corner of rectangular road marking bottom left
-        };
-        // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        BigDecimal[][] gpsDstPts = new BigDecimal[][]{
-                {new BigDecimal(6.603560), new BigDecimal(52.036730d * 1000)}, // Corner of dashed line bottom right
-                {new BigDecimal(6.603227), new BigDecimal(52.036181d * 1000)}, // Corner of white rumble strip top left
-                {new BigDecimal(6.602018), new BigDecimal(52.036769d * 1000)}, // Third lamppost top right
-                {new BigDecimal(6.603638), new BigDecimal(52.036558d * 1000)}, // Corner of rectangular road marking bottom left
-//                {.003560d, .036730d}, // Corner of dashed line bottom right
-//                {.003227d, .036181d}, // Corner of white rumble strip top left
-//                {.002018d, .036769d}, // Third lamppost top right
-//                {.003638d, .036558d}, // Corner of rectangular road marking bottom left
-
-        };
-
-        performConverterTest(gpsSrcPts, gpsDstPts);
-    }
-
-    @Test
-    public void testAlleyCam1Collinearity() {
-        BigDecimal[][] alleyCam1SrcPts = new BigDecimal[][]{
-                {new BigDecimal(65.d), new BigDecimal(10.d)},
-                {new BigDecimal(847.d), new BigDecimal(8.d)},
-                {new BigDecimal(850.d), new BigDecimal(694.d)},
-                {new BigDecimal(63.d), new BigDecimal(696.d)},
-        };
-        // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        BigDecimal[][] alleyCam1DstPts = new BigDecimal[][]{
-                {new BigDecimal(-83.74801820260836), new BigDecimal(42.27796103748212)},
-                {new BigDecimal(-83.74819690478398), new BigDecimal(42.27796396568019)},
-                {new BigDecimal(-83.74813303468181), new BigDecimal(42.78190293489615)},
-                {new BigDecimal(-83.74800177407815), new BigDecimal(42.78187357400434)},
-//                {-.74801820260836d, .27796103748212d},
-//                {-.74819690478398d, .27796396568019d},
-//                {-.74813303468181d, .78190293489615d},
-//                {-.74800177407815d, .78187357400434d},
-        };
-        assertFalse("Destination Points Collinear", CoordinateConverter2.checkCollinearity(alleyCam1DstPts, new BigDecimal(0.00001d)));
-    }
-
-    @Test
-   // @Ignore // not working yet...  Input data suspect
-    public void testAlleyCam1() {
-        // image location --> lat/long
-        // Coords in order topLeft, topRight, bottomRight, bottomLeft (CW from topLeft)
-        BigDecimal[][] alleyCam1SrcPts = new BigDecimal[][]{
-                {new BigDecimal(65.d), new BigDecimal(10.d)},
-                {new BigDecimal(847.d), new BigDecimal(8.d)},
-                {new BigDecimal(850.d), new BigDecimal(694.d)},
-                {new BigDecimal(63.d), new BigDecimal(696.d)},
-        };
-            // longitude, latitude == x,y  -- reversed from traditional specification (but like GeoJSON)
-        BigDecimal[][] alleyCam1DstPts = new BigDecimal[][]{
-                {new BigDecimal(-83.74801820260836), new BigDecimal(42.27796103748212)},
-                {new BigDecimal(-83.74819690478398), new BigDecimal(42.27796396568019)},
-                {new BigDecimal(-83.74813303468181), new BigDecimal(42.78190293489615)},
-                {new BigDecimal(-83.74800177407815), new BigDecimal(42.78187357400434)},
-//                {-.74801820260836d, .27796103748212d},
-//                {-.74819690478398d, .27796396568019d},
-//                {-.74813303468181d, .78190293489615d},
-//                {-.74800177407815d, .78187357400434d},
-        };
-
-        assertFalse("Source Points Collinear", CoordinateConverter2.checkCollinearity(alleyCam1SrcPts));
-        assertFalse("Destination Points Collinear", CoordinateConverter2.checkCollinearity(alleyCam1DstPts, new BigDecimal(0.00001d)));
-
-        performConverterTest(alleyCam1SrcPts, alleyCam1DstPts);
-    }
-
-    protected void performConverterTest(BigDecimal[][] srcPts, BigDecimal[][] dstPts) {
-        performConverterTest(srcPts, dstPts, null, null);
-    }
-
-    protected void performConverterTest(BigDecimal[][] srcPts, BigDecimal[][] dstPts, BigDecimal[][] tstSrc, BigDecimal[][] tstRes) {
-
-        // First, create a converter calibrated by the src & dst points passed in.
-        CoordinateConverter2 converter = new CoordinateConverter2(srcPts, dstPts);
-
-        // verify that that converted correctly converts all the src pts -> dst points
-
-        for (int iter = 0; iter < 2; iter++) {
-            for (int i = 0; i < srcPts.length; i++) {
-                BigDecimal[] src = srcPts[i];
-                BigDecimal[] dst = dstPts[i];
-                BigDecimal[] res = converter.convert(src);
-                if (iter ==1) {
-                    assertTrue("X coord mismatch (" + i + "): expected: " + dst[0] + ", found: " + res[0],
-                            dst[0].compareTo(res[0]) == 0);
-                    assertTrue("Y coord mismatch (" + i + "): expected: " + dst[1] + ", found: " + res[1],
-                            dst[1].compareTo(res[1]) == 0);
-                } else {
-                    log.debug("Iter: {}:{}, X: dst: {}, res:{}, diff: {}", iter, i, dst[0], res[0], dst[0].subtract(res[0]));
-                    log.debug("Iter: {}:{}, Y: dst: {}, res:{}, diff: {}", iter, i, dst[1], res[1], dst[1].subtract(res[1]));
-                }
-            }
-        }
-
-        // If our caller has provided other points, verify that these, too, result in correct values
-
-        if (tstSrc != null) {
-            assertNotNull(tstRes);
-            assertEquals("Src & expected results for test points must be the same length",
-                    tstSrc.length, tstRes.length);
-
-            for (int ti = 0; ti < tstSrc.length; ti++) {
-                BigDecimal[] src = tstSrc[ti];
-                BigDecimal[] dst = tstRes[ti];
-                log.debug("Checking conversion of ({}, {})", src[0], src[1]);
-                BigDecimal[] res = converter.convert(src);
-
-                // Crazy test generation sometimes comes up with odd values.  Scale numbers to reasonable
-                for (int i = 0; i < 2; i++) {
-                    res[i] = res[i].setScale(dst[i].scale(), BigDecimal.ROUND_HALF_UP);
-                }
-                assertTrue("X coord mismatch in test points index: " + ti + ": expected: " + dst[0] + ", found: " + res[0],
-                        dst[0].compareTo(res[0]) == 0);
-                assertTrue("Y coord mismatch in test points index: " + ti + ": expected: " + dst[1] + ", found: " + res[1],
-                        dst[1].compareTo(res[1]) == 0);
-            }
-        }
-    }
-
-    public static BigDecimal[][] generateRandomPoints(int pointCount) {
-        BigDecimal[][] points = new BigDecimal[pointCount][2];
-        for (int i = 0; i < pointCount; i++) {
-            // Generate some random points all over the place to test against...
-            BigDecimal factor1 = new BigDecimal(i % 5 == 0 ? 1d : 10d * (i % 7 == 0 ? pointCount : 1));
-            BigDecimal x = factor1.multiply(new BigDecimal(Math.random()));
-
-            BigDecimal factor2 = new BigDecimal(i % 2 == 0 ? 1d : 130d * (i % 3 == 0 ? pointCount : 1));
-            // Limit craziness of test comparison for randomly generated values...
-            BigDecimal y = factor2.multiply(new BigDecimal(Math.random())).setScale(12, BigDecimal.ROUND_HALF_UP);
-            points[i] = new BigDecimal[]{x, y};
-        }
-        return points;
-    }
-}
